requirements.txt
Flask==3.1.1
markdown==3.8.2
python-frontmatter==1.1.0 